componetize #230
