+++ src/components/app.js
@@ -7,14 +7,70 @@ import AddFold from './add-fold'
 import Editable from './editable'
 import rebassTheme from './rebass-theme'
 
+import EdStore from '../store/ed-store'
 
-class App extends React.Component {
+
+export default class App extends React.Component {
+  constructor (props) {
+    super(props)
+
+    if (!props.initialContent) {
+      throw new Error('Missing props.initialContent')
+    }
+    if (!props.onChange) {
+      throw new Error('Missing props.onChange')
+    }
+    if (!props.onShareUrl) {
+      throw new Error('Missing props.onShareUrl')
+    }
+    if (!props.onShareFile) {
+      throw new Error('Missing props.onShareFile')
+    }
+    if (!props.onRequestCoverUpload) {
+      throw new Error('Missing props.onRequestCoverUpload')
+    }
+
+    const { initialContent
+      , onMount
+      , onChange
+      , onShareFile
+      , onShareUrl
+      , onRequestCoverUpload
+      , onDropFiles
+      , onDropFileOnBlock
+      , onCommandsChanged } = props
+
+    this._store = new EdStore(
+      { initialContent
+      , onMount
+      , onChange
+      , onShareFile
+      , onShareUrl
+      , onRequestCoverUpload
+      , onDropFiles
+      , onDropFileOnBlock
+      , onCommandsChanged
+      }
+    )
+
+    this.routeChange = this._store.routeChange.bind(this._store)
+  }
+  componentDidMount () {
+    this.boundOnDragOver = this.onDragOver.bind(this)
+    window.addEventListener('dragover', this.boundOnDragOver)
+    this.boundOnDrop = this.onDrop.bind(this)
+    window.addEventListener('drop', this.boundOnDrop)
+  }
+  componentWillUnmount () {
+    window.removeEventListener('dragover', this.boundOnDragOver)
+    window.removeEventListener('drop', this.boundOnDrop)
+  }
   getChildContext () {
-    const {store, imgfloConfig} = this.props
+    const {imgfloConfig} = this.props
     return (
       { imgfloConfig: imgfloConfig
       , rebass: rebassTheme
-      , store: store
+      , store: this._store
       }
     )
   }
@@ -29,7 +85,7 @@ class App extends React.Component {
   renderContent () {
     const {initialContent
       , menuBar, menuTip
-      , onChange, onShareFile, onShareUrl
+      , onShareFile, onShareUrl
       , onCommandsChanged, onDropFiles} = this.props
 
     return el('div'
@@ -42,7 +98,7 @@ class App extends React.Component {
       , { initialContent
         , menuBar
         , menuTip
-        , onChange
+        , onChange: this.routeChange
         , onShareFile
         , onShareUrl
         , onCommandsChanged
@@ -50,6 +106,42 @@ class App extends React.Component {
         }
       )
     )
+  }
+  onDragOver (event) {
+    // Listening to window
+    event.preventDefault()
+  }
+  onDrop (event) {
+    // Listening to window, for drops not caught by content
+    event.preventDefault()
+  }
+  // Exposed methods
+  getContent () {
+    return this._store.getContent()
+  }
+  setContent (content) {
+    this._store.setContent(content)
+  }
+  execCommand (commandName) {
+    this._store.execCommand(commandName)
+  }
+  insertPlaceholders (index, count) {
+    return this._store.insertPlaceholders(index, count)
+  }
+  updatePlaceholder (id, metadata) {
+    this._store.updatePlaceholder(id, metadata)
+  }
+  setCoverPreview (id, src) {
+    this._store.setCoverPreview(id, src)
+  }
+  setCover (id, cover) {
+    this._store.setCover(id, cover)
+  }
+  indexOfFold () {
+    return this._store.indexOfFold()
+  }
+  get pm () {
+    return this._store.pm
   }
 }
 App.childContextTypes =
@@ -60,13 +152,12 @@ App.childContextTypes =
 App.propTypes =
   { initialContent: React.PropTypes.array.isRequired
   , onChange: React.PropTypes.func.isRequired
-  , onShareFile: React.PropTypes.func
-  , onShareUrl: React.PropTypes.func
-  , onDropFiles: React.PropTypes.func
+  , onShareFile: React.PropTypes.func.isRequired
+  , onShareUrl: React.PropTypes.func.isRequired
+  , onDropFiles: React.PropTypes.func.isRequired
   , onCommandsChanged: React.PropTypes.func
+  , onRequestCoverUpload: React.PropTypes.func.isRequired
   , menuBar: React.PropTypes.bool
   , menuTip: React.PropTypes.bool
   , imgfloConfig: React.PropTypes.object
-  , store: React.PropTypes.object.isRequired
   }
-export default React.createFactory(App)
