+++ test/plugins/placeholder.js
@@ -1,6 +1,6 @@
 import {expect} from 'chai'
 
-import Ed from '../../src/ed'
+import {mountApp, unmountApp} from '../../src/ed'
 
 
 describe('PluginPlaceholder', function () {
@@ -14,19 +14,18 @@ describe('PluginPlaceholder', function () {
   beforeEach(function (done) {
     mount = document.createElement('div')
     document.body.appendChild(mount)
-    ed = new Ed(
-      { container: mount
-      , initialContent: fixture
+    ed = mountApp(mount
+    , { initialContent: fixture
       , onChange: function () {}
       , onShareUrl: function () {}
       , onShareFile: function () {}
       , onRequestCoverUpload: function () {}
       }
     )
-    ed.on('plugin.placeholder.initialized', done)
+    ed._store.on('plugin.placeholder.initialized', done)
   })
   afterEach(function () {
-    ed.teardown()
+    unmountApp(mount)
     mount.parentNode.removeChild(mount)
   })
